+++ src/main/java/com/example/community/dto/QuestionDTO.java
@@ -6,12 +6,12 @@ import lombok.Data;
 @Data
 public class QuestionDTO {
 
-    private Integer id;     //问题id
+    private Long id;     //问题id
     private String title;       //问题标题
     private String description;     //问题详细描述
     private Long gmtCreate;     //问题创建时间
     private Long gmtModified;   //创建时间戳
-    private Integer creator;    //提问者id
+    private Long creator;    //提问者id
     private Integer commentCount;   //问题评论数
     private Integer viewCount;      //问题浏览数
     private Integer likeCount;      //问题点赞数
